Feat/movie info review dao 병합 전 코드 작성

+++ src/main/java/org/example/model/dto/Genre.java
@@ -1,19 +1,19 @@
 package org.example.model.dto;
 
 public class Genre {
-    private Long genreNo;
+    private int genreNo;
     private String genreName;
 
     public Genre(String genreName) {
         this.genreName = genreName;
     }
 
-    public Genre(Long genreNo, String genreName) {
+    public Genre(int genreNo, String genreName) {
         this.genreNo = genreNo;
         this.genreName = genreName;
     }
 
-    public Long getGenreNo() {
+    public int getGenreNo() {
         return genreNo;
     }
